Moved a lof of funcs to utils and also continued to work on tornado plots
